Move accordion funcs to a util file
